Update convert-user-mailbox-to-shared-mailbox.md
A Global Admin doesn’t have any privilege on-premises, so this information doesn’t make sense. I guess that the point would be covering on how to convert the mailbox in Exchange On-Prem using Powershell. If so, the best thing is to move this Note to the “Hybrid Session” at the end. Another thing is that the command is wrong.