Update set-up-file-storage-and-sharing.md
Added note to clarify use of  the term: "basic team site" which is confusing as per https://github.com/MicrosoftDocs/OfficeDocs-O365ITPro/issues/170 as the site is there but hidden and should actually not be mentioned at all. The whole article may need revision along with https://docs.microsoft.com/office365/admin/setup/cutomize-team-site/ which mentions a "classic team site" which is also not accessible to most users and is confusing as, like the "basic team site" it predates, and is not related to Microsoft Teams at all.