Updates related to logging into converted mailbox
It was not explicitly clear that the user's old username / password actually WILL be able to log into the converted shared mailbox after conversion. Added information on resetting the password (if user was deleted) and what other things need to be done if this is an employee that has left the organization.