Update restore-deleted-group.md
The descriptions of the permission were wrong, so I fixed it.

When restoring the group using EAC, the user must be an administrator.
When restoring the group using AzureADPreview command, if the user is an administrator, the user can restore all deleted groups. Even if the user is not an administrator, the user can restore the group which the user owns.