Altering TIP regarding change of user UPN
Line 246
The tip states that the change of a user's UPN will have no bearing on their email address. Suggest that it's specified that no change will happen to the MOERA email address, or that the tip it is removed altogether since this is false.
Altering a user's UPN will currently trigger a recalculation of the proxyAddresses, and it will include the new UPN as either a secondary or primary SMTP depending on additional circumstances.