[DO NOT MERGE] Edits to External participants in Yammer
Made requested edits to the article.

Added an intro sentence to the body to describe why an admin would want this list. I leveraged it from add-external-participants.md,  but the author should verify that it accurately describes the reason.

Moved content in step 2 to a note above the image, and indented info and the image that's directly part of step 2.

Minor edits for clarity.